requirements.txt
Flask==2.2.5
requests==2.31.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
python-dotenv==0.21.1